[Segment Cache] Add refresh URL to reused default segments (#84627)
When segment is reused in the "default" parallel route slot, it is not
part of the new page. So if the page is refreshed, we must fetch its
data from the previous URL, not the new URL. To track this we add the
previous URL to a special field in the FlightRouterState.

Currently this field is added by
addRefreshMarkerToActiveParallelSegments; this updates the
"ppr-navigations" module to add it during the main traversal used by
navigations, so we can eventually get rid
addRefreshMarkerToActiveParallelSegments.

Note that we don't need to add the marker to _all_ page segments, just
the ones that are not part of the new page.