[Cache Components] Disable static indicator (#84639)
When `experimental.cacheComponents` is enabled, the static indicator in
the Next.js DevTools is now disabled.

With Cache Components enabled, the binary static/dynamic state is no
longer accurate since routes are typically partially static and dynamic
(unless fully opted into dynamic rendering via a Suspense boundary above
the body, which is an edge case).

To avoid confusion, we're removing it for now, when the feature flag is
enabled. In the future, we will likely provide better tools that will
allow users to understand which parts of a page are prerendered and
which parts are dynamic.

In addition, this PR also improves the behavior for fully static App
Router pages (without Cache Components). Previously, we would show
"Dynamic" as default, and only flipped to "Static" when the request was
finished. Now, we're showing a spinner while the page is loading.

![static
indicator](https://github.com/user-attachments/assets/c720975d-932c-48b0-b705-7c408fdcf51d)

closes NAR-428