[turbopack] Emit a warning when there are too many matches from a FileSourceReference (#84701)
Previously we would `eprintln!` in this case, but that is missing key source information that would help with debugging.


Example warning (i hacked the limit to be lower to make this less tedious to repro):

![image.png](https://app.graphite.dev/user-attachments/assets/084b9b27-2558-4131-8f31-e54e3c188359.png)

If there are more than 10K matches we will issue this warning.


A tricky thing is that because we issue this warning during the module graph construction, but `resolve_reference` won't be called again for a FileSource reference during chunking the issue actually wouldn't get emitted.  The fix is to drop the `take_issues` call inside of `whole_app_module_graphs` in production builds.