[ci]: set NEXT_TEST_VERSION to release tag for deploy tests (#84840)
When deploy tests run on release branches, they end up using "canary"
rather than the version associated with the release. This means that the
test files might be misaligned with the underlying versions they are
testing.

We are already doing the work of collecting the next version, we just
weren't plumbing it into `NEXT_TEST_VERSION`.