[turbopack] Allow withRspack to work even if you already have NEXT_RSPACK set (#84987)
Fix an issue where if a user set `NEXT_RSPACK` as an environment variable then `next` would think it was using rspack but `withRspack` would think it was using `--webpack`.   The issue is that the idempotency check was too strict.

This partially reverts https://github.com/vercel/next.js/pull/84885