Turbopack: Track errored tasks as dependency when using untracked() (#84914)
### What?

The pattern `some_operation().untracked().await?` is dangerous as it propages the error to the parent task.

An untracked read accepts an incorrectness due to eventual consistency during an update. This might also cause eventual consistent errors.

If these errors are propagates to the parent task, these error will not easily go away, as the read was untrack and therefore the error doesn't resolve automatically when the `some_operation` task computes again.

This change changes to the default of untracked read to be only untracked for the value, but track the dependency in case of an error.