Add rendered search to router state (#84983)
The "rendered search" represents the search params that are observed on
the server, i.e. the ones passed as the `searchParams` prop to page
segments. They may differ from the search params in the browser's URL
bar, if the server performed a rewrite. This case corresponds to the
x-nextjs-rewritten-query header.

If a page segment is marked with `"use client"`, we can use this value
to compute the search params on the client instead of having to fetch
them from the server. This part is implemented in #84883.