Update next-lint-to-eslint-cli to support `FlatCompat.config` (#85026)
ESLint `FlatCompat` can be used in two ways:

```js
compat.extends('next/core-web-vitals', 'next/typescript')
```

```js
compat.config({
  extends: ['next/core-web-vitals', 'next/typescript'],
})
```

Although using `.extends()` is recommended for migration, and the ESLint
migration tool also uses that, it is possible to use `.config()`, so
handle them as well for the migration.

Confirmed on local apps as well.