Add comment that we expect the function passed to bind to be anonymous (#85070)
See https://github.com/facebook/react/pull/34911

This is a weird one but we need some consistent name to ignore these
because otherwise they show up as I/O.

The simplest is to simply not assign a name to the function being
passed. It also can't be auto-assigned by a compiler. Because whatever
name the function is given is the name given to the AsyncResource.