Await initial Flight response before hydrating (#85124)
When hydrating on the client, we can await the promise for the initial
Flight response before we call hydrateRoot. Previously we would unwrap
this promise inside a component with `use`, which caused it to be
treated as a suspense "dependency" in the Suspense DevTools panel. This
is unnecessary noise because the latency of the initial Flight data is
already encoded into the Flight data itself, as opposed to the promise
that resolves to the Flight data.

This removes an unnecessary piece of debug information from the Suspense
DevTools panel.