[Turbopack] dedupe build errors (#85062)
When we run the RSC validation transform, we might take multiple passes
on the same file if the edge runtime is used: once at the RSC layer,
another in the edge RSC layer. This means we'll have reported the error
twice for the same file. This dedupes errors by tracking the seen,
formatted errors and not re-showing them if they've already been
tracked.