[Cache Components] set environment labels if caches are disabled (#85129)
If we're bypassing caches, we still want to provide some basic
environment labelling and correct "suspended by" labels for Suspense
Devtools. To achieve this, we do a simpler 2-task render, but still use
`stagedRendering` and install all the `asyncApiPromises` stuff (because
it's needed for "suspended by"). Note that because we only have 2 tasks
and don't warm caches, we can only do `Prerender` and `Server`, and
caches are marked as IO. This isn't ideal but it's the best we can do
under the circumstances.