Make params and searchParams new Promises (#85158)
The problem is that without a new Promise, when all the awaits are in
third party, we'd keep searching down to the inner most which shows
`internal params`. We could maybe make the real value of those include
all params. But we can also just not dedupe to show the real value.