Assign name to hydration stream (#85179)
This allows React DevTools to show a name for this which doesn't have a
url like the fetches.

Also, don't set startTime for client resume fetch. The start time for
this data didn't start on the server as an MPA nav. The start of the
fetch actually is a waterfall and should be shown as such.