docs: remove inaccuracies from use cache: private (#85425)
There are a lot of things in here that aren't quite fully fleshed out
yet (eg runtime prefetching) as well as some inaccuracies (eg `use
cache: private` is not required for runtime prefetching, it's just
another thing that _can_ be runtime prefetched).

This trims the doc down until we can focus on proper guidance.