Generalize Segment Cache fallback implementation (#84652)
This is a stack of refactors related to the internal map type used to
store Segment Cache entries on the client. Refer to each commit message
for full details.

The main motivation is to generalize the "fallback entry" mechanism used
for interception routes to more kinds of params. Currently this is only
used for interception routes and search params, but in the future we
will use it for route params, too.