fix: staleTimes.static should consistently enforce a 30s minimum (#85479)
When the tree response sends back a static staleTime, we had handling to
ensure that it was at minimum 30s, because anything less would mean we'd
be discarding static cache entries more frequently than necessary.
Currently a value of 0 would result in a prefetch ping loop, because the
cache value would be immediately invalidated.

This updates our config validation to warn about values lower than 30s
and ensures the `getStaleTimeMs` function applies to the value read from
env, too.