Refactor GTM implementation to support google tag gateway (#81011)
### What?

This PR refactors the Google Tag Manager implementation on
`@next/third-parties`, it rewrites the implementation of the script url,
params and types to allow `gtmId` to be optional when a script URL is
provided.

### Why?

Google released [Google Tag Gateway for
Advertisers](https://developers.google.com/tag-platform/tag-manager/gateway/setup-guide?setup=manual)
which allows first-party tracking on GTM. When running GTM using the
current implementation, it doesn't work because the `?id=` is hardcoded
into the implementation as it doesn't consider the possibility of tag
gateway or any other setup where the ID is already part of the url.

### How?

- Expanded the types
- Refactored the initialisation of the script URL
- Updated documentation relevant to GTM

Closes #81009

Co-authored-by: JJ Kasper <jj@jjsweb.site>