[turbopack] Move generation of `cacheLife` types out of the webpack plugin and into the dev bundler directly (#85539)
Ensure that users configuring custom `cacheLife` configs with Typescript get support with turbopack

This follows the same pattern as typed-routes, so generation can happen during `next type-gen` and also via the dev bundler.

Also this fixes a subtle bug in the timestamp formatting where we believed months had 30 or 31 _weeks_ instead of 30 or 31 _days_s.

Beyond the unit tests, i also tried out adding a `cacheLife` config to the `with-turbopack` example and observed that the types were generated and vscode was able to autocomplete my custom config type.

Fixes NAR-458