Turbopack: chore: Remove some dead MagicAny serialization code from turbo_tasks::value (#85577)
Noticed this was dead while I was looking through how we might integrate bincode.

We only ever use `any_serialization` and not `magic_serialization`. In theory, `magic_serialization` is slightly better because it can include the type name in the error message.

If we go through with the bincode stuff, I'll replace `any_serialization` and the `MagicAny` serialization with something that doesn't depend on `erased-serde`.