[turbopack] Support relative paths in turbopack source maps. (#85146)
### What

Add an option to turbopack to render sourcemap relative file paths in source maps

For now this is unused since it requires some support in next to be added in #85576

### Why
This will be useful for server source maps because it enables node and other runtimes to automatically format file paths to be working directory relative when printing errors to the console which many terminal emulators support to make clickable.


### TIL

My initial approach for this was to set the `sourceRoot` option to avoid prepending the same string onto nearly every source uri.  However, the spec on this sourcemap feature is garbage.  Intuitively i thought it worked like `<base href=...>` where relative uris would 'resolve' against it, but no, it is just a string concatenation feature. sigh.