Don't invalidate hot reloader excessively during dev server boot (#85732)
Suspecting that this fixes (or reduces) the "Watchpack Error (watcher):
Error: EMFILE:" issues e.g.
https://github.com/vercel/next.js/actions/runs/19028750529/job/54337767548?pr=85637

I noticed that all the Watchpack errors in CI happened in tests with a
tsconfig (or ones where a tsconfig was created). When watchpack starts
watching, an initial `aggregated` event is set. During that event we
used to consider all files as changed because we hadn't seen them
before. This also meant we considered the tsconfig as changed and
invalidated the hot-reloader.

This is excessive and likely hits some hard to debug race conditions.
Now we check if the file was actually changed since we started watching
by comparing its change time to the time when we started watching. Each
file modified before we started watching is no longer considered
changed.